extracted abstract VolatileUser out of GuestUser
